Fix: CSVDatabase is now generic and main creates the db.
