+++ Chirp.CLI/Program.cs
@@ -1,8 +1,10 @@
-﻿using System.Globalization;
-using CsvHelper;
+﻿using System.ComponentModel.Design;
+using System.Globalization;
 using SimpleDb;
 
 string timeFormat = "MM/dd/yy HH:mm:ss";
+
+IDatabaseRepository<Cheep> db = new CSVDatabase<Cheep>();
 
 if (args.Length < 1) 
 {  
@@ -13,7 +15,7 @@ if (args.Length < 1)
 
 if (args[0] == "read")
 {
-    Read();
+    db.Read();
 } 
 else if (args[0]  == "cheep")
 {
@@ -21,5 +23,24 @@ else if (args[0]  == "cheep")
     {
         Console.WriteLine("You are missing a message to cheep :(");
     }
-    Store(args[1]);
+
+    db.Store(new Cheep(Environment.UserName, args[1], 
+             FromDateTimeToUnix(DateTime.Now.ToString(timeFormat, CultureInfo.InvariantCulture))));
 }
+
+    long FromDateTimeToUnix(string dateTimeStamp)
+    {
+        DateTime parsedTime = DateTime.Parse(dateTimeStamp, CultureInfo.InvariantCulture);
+        return new DateTimeOffset(parsedTime).ToUnixTimeSeconds();
+    }
+
+    //For printing purposes.
+    /*string FromUnixTimeToDateTime(long timestamp)
+        {
+            DateTimeOffset dto = DateTimeOffset.FromUnixTimeSeconds(timestamp);
+            string correctFormatTimestamp = dto.ToLocalTime().ToString(timeFormat, CultureInfo.InvariantCulture);
+            return correctFormatTimestamp;
+        }
+    */
+
+record Cheep(string Author, string Message, long Timestamp);
