+++ Chirp.CLI/chirp_cli_db.csv
@@ -1,8 +1,3 @@
 Author,Message,Timestamp
-ropf,"Hello, BDSA students!",1690891760
-adho,"Welcome to the course!",1690978778
-adho,"I hope you had a good summer.",1690979858
-ropf,"Cheeping cheeps on Chirp :)",1690981487
-OmarK,"Is it working?",1724854543
-asger,"testing",1725458609
-ander,testing,1725461586
+ander,TEST fdsa asdf sadf asdf 123,1725970231
+ander,TEST fdsa afdsfds,1725970812
