+++ SimpleDb/IDatabaseRepository.cs
@@ -1,6 +1,6 @@
 namespace SimpleDb;
 
-interface IDatabaseRepository<T>
+public interface IDatabaseRepository<T>
 {
     public IEnumerable<T> Read(int? limit = null);
     public void Store(T record);
